+++ src/master/master_impl.h
@@ -35,7 +35,7 @@ struct AgentInfo {
     Agent_Stub* stub;
     int32_t alive_timestamp;
     std::set<int64_t> running_tasks;
-    int32_t version;
+    int64_t version;
 };
 
 struct AgentLoad{
